Added styling to form agreements and removed unused view
