+++ assets/global-classes.less
@@ -202,7 +202,6 @@ textarea {
   align-items: center;
   position: relative;
   margin: 0 !important;
-  padding-left: 3rem;
   cursor: pointer;
   font-size: 1rem;
   -webkit-user-select: none;
@@ -228,6 +227,14 @@ textarea {
   background-color: @tangerine;
 }
 
+.checkmark {
+  border: 2px solid #888;
+  border-radius: 6px;
+  height: 30px;
+  width: 30px;
+  margin: 0 auto;
+  background-color: transparent;
+}
 .checkmark:after {
   content: "";
   position: absolute;
