--- src/views/SignUpAgreements.Vue.vue
@@ -1,21 +1,0 @@
-<template>
-  <div>
-    <h2>Agreements</h2>
-    <form-agreements />
-  </div>
-</template>
-
-<script>
-import FormAgreements from '../components/Forms/FormAgreements.vue';
-
-export default {
-  name: 'SignUpAgreements.Vue',
-  components: {
-    FormAgreements,
-  },
-};
-</script>
-
-<style scoped>
-
-</style>
